Fix class not found issue when using default JDBC jar
